genai/requirements.txt

streamlit
PyMuPDF
sumy
transformers==4.30.2
torch
sentence-transformers
farm-haystack
fuzzywuzzy
python-Levenshtein
